pyproject.toml
[tool.poetry]
name = "tefnut"
version = "0.4.0"
description = "Tefnut is a control system to manage central humidificatior."
authors = ["Marc-Olivier Arsenault <marcolivier.arsenault@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
dynaconf = "*"
python-logging-loki = "*"
influxdb-client = "*"
logging-json = "*"
pyecobee = "*"
psutil = "*"
pytest-cov = "*"
requests-mock = "*"
gitpython = "*"
flask = "*"
flask-login = "*"
pytest-mock = "*"
gunicorn = "*"
flask-apscheduler = "*"
isort = "^5.11.4"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
